# old_src/DONE_admin/application/adapters/in/docs/swagger.json
{
  "components": {
    "schemas": {
      "CreateProductBody": {
        "properties": {
          "category": {
            "enum": ["Lanches", "Bebidas", "Acompanhamentos", "Sobremesas"],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          }
        },
        "required": ["name", "category", "price"],
        "type": "object"
      },
      "DeleteProductBody": {
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": ["id"],
        "type": "object"
      },
      "Product": {
        "properties": {
          "category": {
            "example": "Lanches",
            "type": "string"
          },
          "id": {
            "example": "60d21b4667d0d8992e610c85",
            "type": "string"
          },
          "name": {
            "example": "Hambúrguer",
            "type": "string"
          },
          "price": {
            "example": 25.5,
            "type": "number"
          }
        },
        "type": "object"
      },
      "UpdateProductBody": {
        "properties": {
          "category": {
            "enum": ["Lanches", "Bebidas", "Acompanhamentos", "Sobremesas"],
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          }
        },
        "required": ["id", "name", "category", "price"],
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Admin Products API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/products": {
      "delete": {
        "description": "Delete a product from the inventory.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteProductBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            },
            "description": "Product deleted successfully"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Delete a product",
        "tags": ["Products"]
      },
      "patch": {
        "description": "Update details of an existing product.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            },
            "description": "Product updated successfully"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Update an existing product",
        "tags": ["Products"]
      },
      "post": {
        "description": "Create a new product in the inventory.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            },
            "description": "Product created successfully"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Create a new product",
        "tags": ["Products"]
      }
    },
    "/admin/products/{category}": {
      "get": {
        "description": "Retrieve products by category.",
        "parameters": [
          {
            "description": "Category of the products",
            "in": "path",
            "name": "category",
            "required": true,
            "schema": {
              "enum": ["Lanches", "Bebidas", "Acompanhamentos", "Sobremesas"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  },
                  "type": "array"
                }
              }
            },
            "description": "A list of products"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "example": "Categoria inválida",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Invalid category"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Get products by category",
        "tags": ["Products"]
      }
    }
  },
  "tags": [
    {
      "description": "Endpoints relacionados aos produtos",
      "name": "Products"
    }
  ]
}